Feature a user-input based grid maker and drawing capabilities
The initial grid of 50x50 squares can be modified by prompting the user
upon the push of a button to the enter the number of squares per side,
which instantly generates a new grid. Note that the maximum limit is set
to 100 squares due to performance and rendering issues.
Next, each square is highlighted by mouse hover, and permanently
filled in with a dark color if the mouse moves over it, between pressing
down and releasing it.